% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{occupations}
\alias{occupations}
\title{Occupations}
\format{
A data frame with 3948 rows and 2 columns:
\describe{
\item{baldesig}{Ballot designation as it appears in the CEDA dataset}
\item{hand_coded}{A hand-coded occupation classification (for a random subset)}
}
}
\usage{
data(occupations)
}
\description{
This dataset contains 3,948 ballot designations from municipal elections in California.
A random subset are hand-labeled as either "Working Class" or "Not Working Class" occupations.
}
\references{
California Elections Data Archive (CEDA). https://hdl.handle.net/10211.3/210187
}
\keyword{datasets}
