% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{occupations_examples}
\alias{occupations_examples}
\title{Labelled Occupations}
\format{
A data frame with 9 rows and 2 columns:
\describe{
\item{text}{The text of the ballot designation}
\item{label}{The hand-coded label (Working Class, Not Working Class, NA)}
}
}
\usage{
data(occupations_examples)
}
\description{
This dataset contains 9 example occupations
along with a classification. These can be used as few-shot
examples for classifying occupations in the \code{occupations} dataset.
}
\references{
California Elections Data Archive (CEDA). https://hdl.handle.net/10211.3/210187
}
\keyword{datasets}
