% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{scotus_tweets}
\alias{scotus_tweets}
\title{Tweets About The Supreme Court of the United States}
\format{
A data frame with 945 rows and 5 columns:
\describe{
\item{tweet_id}{A unique ID}
\item{text}{The text of the tweet}
\item{case}{An identifier denoting which Supreme Court ruling the tweet was collected after.}
\item{expert1, expert2, expert3}{Hand-coded sentiment score (-1 = negative, 0 = neutral, 1 = positive)}
}
}
\usage{
data(scotus_tweets)
}
\description{
This dataset contains 945 tweets referencing the US Supreme Court.
Roughly half were collected on June 4, 2018 following the \emph{Masterpiece Cakeshop}
ruling, and the other half were collected on July 9, 2020 following the
Court's concurrently released opinions in \emph{Trump v. Mazars} and \emph{Trump v. Vance}.
Each tweet includes three independent human-coded sentiment scores (-1 to +1).
}
\details{
CONTENT WARNING: These texts come from social media, and many contain explicit
or offensive language.
}
\references{
Ornstein et al. (2024). "How To Train Your Stochastic Parrot"
}
\keyword{datasets}
