% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{scotus_tweets_examples}
\alias{scotus_tweets_examples}
\title{Labelled Example Tweets About The Supreme Court of the United States}
\format{
A data frame with 12 rows and 4 columns:
\describe{
\item{tweet_id}{A unique ID for each tweet}
\item{text}{The text of the tweet}
\item{case}{The case referenced in the tweet (Masterpiece Cakeshop or Trump v. Mazars)}
\item{label}{The "true" label (Positive, Negative, or Neutral)}
}
}
\usage{
data(scotus_tweets_examples)
}
\description{
This dataset contains 12 example tweets referencing the Supreme Court
along with a sentiment label. These can be used as few-shot prompt
examples for classifying tweets in the \code{scotus_tweets} dataset.
}
\references{
Ornstein et al. (2023). "How To Train Your Stochastic Parrot"
}
\keyword{datasets}
